done too ig

+++ customer.php
@@ -19,8 +19,34 @@
         </div>
 
         <div class="bar">
+            <div id="mySidenav" class="sidenav">
+                <a href="javascript:void(0)" class="closebtn" onclick="closeNav()">&times;</a>
+                <a href= "actor.php">actor</a>
+                <a href= "address.php">address</a>
+                <a href= "category.php">category</a>
+                <a href= "city.php">city</a>
+                <a href= "country.php">country</a>
+                <a href= "customer.php">customer</a>
+                <a href= "film_actor.php">film_actor</a>
+                <a href= "film_category.php">film_category</a>
+                <a href= "film_text.php">film_text</a>
+                <a href= "film.php">film</a>
+                <a href= "inventory.php">inventory</a>
+                <a href= "language.php">language</a>
+                <a href= "payment.php">payment</a>
+                <a href= "rental.php">rental</a>
+                <a href= "staff.php">staff</a>
+                <a href= "store.php">store</a>
+                <a href= "payment&customer&staff.php">payment&customer&staff</a>
+                <a href= "film&film_text&film_category&category&language.php">film&film_text&film_category&category&language</a>
+                <a href= "country&city&address.php">country&city&address</a>
+            </div>
+
+            <div class="menu">
+                <span style="font-size:25px;cursor:pointer;color:white; text-align:left; display:inline;" onclick="openNav()">&#9776; MENU</span>
+            </div>
             <form action= "" method= "POST" style= 'display: inline;'>
-                <input type ="text" name= "search" placeholder="SEARCH BY ID" style= "border-radius: 5px;">
+                <input type ="text" name= "search" placeholder="SEARCH BY CUSTOMER ID" style= "border-radius: 5px;">
             </form>
             <form action= "" method= "POST" style= 'display: inline;'>
                 <input type = "submit" name= "reset" value= "RESET" class="button"> 
@@ -89,36 +115,40 @@
 
                     <?php
                     if(isset($_POST['insert'])){
-                        if(!empty($_POST['customerid'])&& !empty($_POST['storeid'])&& !empty($_POST['firstname'])&& !empty($_POST['lastname'])&& !empty($_POST['email'])&& !empty($_POST['addressid'])&& !empty($_POST['active'])){
-                            $customerid= $_POST['customerid'];
-                            $query= "SELECT customer_id FROM customer WHERE customer_id = $customerid;";
-                            $result= mysqli_query($conn,$query);
-                            
-                            if(mysqli_num_rows($result)==0){
+                        if(!empty($_POST['customerid'])&& !empty($_POST['storeid'])&& !empty($_POST['firstname'])&& !empty($_POST['lastname'])&& !empty($_POST['email'])&& !empty($_POST['addressid'])&& ($_POST['active']==1 || $_POST['active']==0)){
+                            if((is_numeric($_POST['customerid'])) && ($_POST['customerid'] > 0)){
                                 $customerid= $_POST['customerid'];
-                                $storeid= $_POST['storeid'];
-                                $firstname= $_POST['firstname'];
-                                $lastname= $_POST['lastname'];
-                                $email= $_POST['email'];
-                                $addressid= $_POST['addressid'];
-                                $active= $_POST['active'];
-                                $createdate= date('Y-m-d H:i:s');
-                                $insert = "INSERT INTO customer VALUES('$customerid','$storeid','$firstname','$lastname','$email','$addressid','$active','$createdate','$createdate');";
-                                $result = mysqli_query($conn,$insert);
-                                if ($result) {
-                                    echo '<script> alert("DATA INSERTED SUCCESSFULLY!")</script>';
-                                }
-                                else
-                                    echo '<script> alert("PREVIOUS INSERT FAILED! CHECK IF THERE WERE MISTAKES MADE WHEN INSERTING")</script>'; 
-                                
-                                echo("<meta http-equiv='refresh' content='1'>");
-                            }else{
-                                echo '<script> alert("PREVIOUS INSERT FAILED! CHECK IF THERE WERE MISTAKES MADE WHEN INSERTING")</script>';
-                            }
+                                $query= "SELECT customer_id FROM customer WHERE customer_id = $customerid;";
+                                $result= mysqli_query($conn,$query);
                             
+                                if(mysqli_num_rows($result)==0){
+                                    $customerid= $_POST['customerid'];
+                                    $storeid= $_POST['storeid'];
+                                    $firstname= $_POST['firstname'];
+                                    $lastname= $_POST['lastname'];
+                                    $email= $_POST['email'];
+                                    $addressid= $_POST['addressid'];
+                                    $active= $_POST['active'];
+                                    $createdate= date('Y-m-d H:i:s');
+                                    $firstname= strtoupper($firstname);
+                                    $lastname= strtoupper($lastname);
+                                    $insert = "INSERT INTO customer VALUES('$customerid','$storeid','$firstname','$lastname','$email','$addressid','$active','$createdate','$createdate');";
+                                    $result = mysqli_query($conn,$insert);
+                                    if ($result) {
+                                        echo '<script> alert("DATA INSERTED SUCCESSFULLY!")</script>';
+                                    }
+                                    else
+                                        echo '<script> alert("PREVIOUS INSERT FAILED! STORE ID OR ADDRESS ID ENTERED DOES NOT EXIST, PLEASE CHECK THE STORE AND ADDRESS TABLES FOR AN EXISTING STORE ID AND ADDRESS ID RESPECTIVELY")</script>';
                                 
+                                    echo("<meta http-equiv='refresh' content='1'>");
+                                }else{
+                                    echo '<script> alert("PREVIOUS INSERT FAILED! CUSTOMER ID ENTERED ALREADY EXISTS")</script>';
+                                }
+                            }else{
+                                echo '<script> alert("PREVIOUS INSERT FAILED! INVALID CUSTOMER ID ENTERED")</script>';
+                            }
                         }else{
-                            echo '<script> alert("PREVIOUS INSERT FAILED, PLEASE FILL ALL FIELDS!")</script>';
+                            echo '<script> alert("PREVIOUS INSERT FAILED! PLEASE FILL ALL FIELDS AND MAKE SURE ACTIVE HOLDS EITHER 0 OR 1")</script>';
                         }
                     }
                     ?>
@@ -148,34 +178,39 @@
 
                     <?php
                     if(isset($_POST['update'])){
-                        if(!empty($_POST['customerid'])&& !empty($_POST['storeid'])&& !empty($_POST['firstname'])&& !empty($_POST['lastname'])&& !empty($_POST['email'])&& !empty($_POST['addressid'])&& !empty($_POST['active'])){
-                            $customerid= $_POST['customerid'];
-                            $query= "SELECT customer_id FROM customer WHERE customer_id = $customerid;";
-                            $result= mysqli_query($conn,$query);
-
-                            if(mysqli_num_rows($result)==1){
+                        if(!empty($_POST['customerid'])&& !empty($_POST['storeid'])&& !empty($_POST['firstname'])&& !empty($_POST['lastname'])&& !empty($_POST['email'])&& !empty($_POST['addressid'])&& ($_POST['active']==1 || $_POST['active']==0)){
+                            if((is_numeric($_POST['customerid'])) && ($_POST['customerid'] > 0)){
                                 $customerid= $_POST['customerid'];
-                                $storeid= $_POST['storeid'];
-                                $firstname= $_POST['firstname'];
-                                $lastname= $_POST['lastname'];
-                                $email= $_POST['email'];
-                                $addressid= $_POST['addressid'];
-                                $active= $_POST['active'];
-                                $lastupdate= date('Y-m-d H:i:s');
-                                $update = "UPDATE customer SET store_id='$storeid',first_name='$firstname',last_name='$lastname',email='$email',address_id='$addressid',active='$active',last_update='$lastupdate' WHERE customer_id='$customerid';";
-                                $result = mysqli_query($conn,$update);
-                                if($result)
-                                    echo '<script> alert("DATA UPDATED SUCCESSFULLY!")</script>';
-                                else
-                                    echo '<script> alert("PREVIOUS UPDATE FAILED! CHECK IF THERE WERE MISTAKES MADE WHEN UPDATING")</script>';  
+                                $query= "SELECT customer_id FROM customer WHERE customer_id = $customerid;";
+                                $result= mysqli_query($conn,$query);
+
+                                if(mysqli_num_rows($result)==1){
+                                    $customerid= $_POST['customerid'];
+                                    $storeid= $_POST['storeid'];
+                                    $firstname= $_POST['firstname'];
+                                    $lastname= $_POST['lastname'];
+                                    $email= $_POST['email'];
+                                    $addressid= $_POST['addressid'];
+                                    $active= $_POST['active'];
+                                    $lastupdate= date('Y-m-d H:i:s');
+                                    $firstname= strtoupper($firstname);
+                                    $lastname= strtoupper($lastname);
+                                    $update = "UPDATE customer SET store_id='$storeid',first_name='$firstname',last_name='$lastname',email='$email',address_id='$addressid',active='$active',last_update='$lastupdate' WHERE customer_id='$customerid';";
+                                    $result = mysqli_query($conn,$update);
+                                    if($result)
+                                        echo '<script> alert("DATA UPDATED SUCCESSFULLY!")</script>';
+                                    else
+                                        echo '<script> alert("PREVIOUS UPDATE FAILED! STORE ID OR ADDRESS ID ENTERED DOES NOT EXIST, PLEASE CHECK THE STORE AND ADDRESS TABLES FOR AN EXISTING STORE ID AND ADDRESS ID RESPECTIVELY")</script>'; 
                                 
-                                echo("<meta http-equiv='refresh' content='1'>");
+                                    echo("<meta http-equiv='refresh' content='1'>");
+                                }else{
+                                    echo '<script> alert("PREVIOUS UPDATE FAILED! CUSTOMER ID ENTERED DOES NOT EXIST")</script>';
+                                }   
                             }else{
-                                echo '<script> alert("PREVIOUS UPDATE FAILED! CHECK IF THERE WERE MISTAKES MADE WHEN UPDATING")</script>';
+                                echo '<script> alert("PREVIOUS UPDATE FAILED! INVALID CUSTOMER ID ENTERED")</script>';
                             }
-                                
                         }else{
-                            echo '<script> alert("PREVIOUS UPDATE FAILED, PLEASE FILL ALL FIELDS!")</script>';
+                            echo '<script> alert("PREVIOUS UPDATE FAILED! PLEASE FILL ALL FIELDS AND MAKE SURE ACTIVE HOLDS EITHER 0 OR 1")</script>';
                         }
                     }
                     ?>
@@ -196,27 +231,30 @@
             <?php
                     if(isset($_POST['delete'])){
                         if(!empty($_POST['customerid'])){
-                            $customerid= $_POST['customerid'];
-                            $query= "SELECT customer_id FROM customer WHERE customer_id = $customerid;";
-                            $result= mysqli_query($conn,$query);
-
-                            if(mysqli_num_rows($result)==1){
+                            if((is_numeric($_POST['customerid'])) && ($_POST['customerid'] > 0)){
                                 $customerid= $_POST['customerid'];
-                                $delete = "DELETE FROM customer WHERE customer_id='$customerid';";
-                                $result = mysqli_query($conn,$delete); 
-                                if($result){
-                                    echo '<script> alert("ROW DELETED SUCCESSFULLY!")</script>';
+                                $query= "SELECT customer_id FROM customer WHERE customer_id = $customerid;";
+                                $result= mysqli_query($conn,$query);
+
+                                if(mysqli_num_rows($result)==1){
+                                    $customerid= $_POST['customerid'];
+                                    $delete = "DELETE FROM customer WHERE customer_id='$customerid';";
+                                    $result = mysqli_query($conn,$delete); 
+                                    if($result){
+                                        echo '<script> alert("ROW DELETED SUCCESSFULLY!")</script>';
+                                    }else{
+                                        echo '<script> alert("DELETE FAILED! YOU ARE NOT ALLOWED TO DELETE THIS ROW")</script>';
+                                    } 
+                                
+                                    echo("<meta http-equiv='refresh' content='1'>");
                                 }else{
-                                    echo '<script> alert("DELETE FAILED! YOU ARE NOT ALLOWED TO DELETE THIS ROW")</script>';
-                                } 
-                                
-                                echo("<meta http-equiv='refresh' content='1'>");
+                                    echo '<script> alert("PREVIOUS DELETE FAILED! CUSTOMER ID ENTERED DOES NOT EXIST")</script>';
+                                }
                             }else{
-                                echo '<script> alert("PREVIOUS DELETE FAILED! YOU ARE NOT ALLOWED TO DELETE THIS ROW")</script>';
+                                echo '<script> alert("PREVIOUS DELETE FAILED! INVALID CUSTOMER ID ENTERED")</script>';
                             }
-                                
                         }else{
-                            echo '<script> alert("PREVIOUS DELETE FAILED, PLEASE FILL ALL FIELDS!")</script>';
+                            echo '<script> alert("PREVIOUS DELETE FAILED! PLEASE FILL ALL FIELDS")</script>';
                         }
                     }
                     ?>
@@ -253,6 +291,14 @@
             function(){
                 document.querySelector('.delete').style.display= 'none';
             });
+            
+            function openNav() {
+                document.getElementById("mySidenav").style.width = "620px";
+            }
+
+            function closeNav() {
+                document.getElementById("mySidenav").style.width = "0";
+            }
         </script>
     </body>
 </html>
